+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 * Job summary: error column does not display errors (UIDATIMP-1590)
 * Handle empty marc mapping label (UIDATIMP-1592)
 * Update style of Incoming Record button to match other buttons (UIDATIMP-1594)
+* Not able to use system generated Match profiles (UIDATIMP-1521)
 
 ## [7.0.5](https://github.com/folio-org/ui-data-import/tree/v7.0.5) (2023-12-06)